Move padding to hero
